requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
gunicorn==21.2.0
sqlalchemy==2.0.29
psycopg2-binary==2.9.9
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
boto3==1.34.69
pydantic[email]==2.6.4
python-multipart==0.0.9
structlog==24.1.0
prometheus-fastapi-instrumentator==6.1.0
